Added functions for adding mp/hp to a Char. Fixed items with potential getting a proper swirly when on the ground. Moved mob packets to a seperate class. Fixed many use items. Buff potions give you the stat, but don't display the buff icon. Fixed mobs not attacking. Made a start on mob skills, but I can't get the icon to display (like damage reflect). Fixed mob movement parsing. Added a type for inventory operations. Also removed redudant parameters to the packet's function. Fixed mobs not going back to their initial position after respawning. Added money being based off of the mob's exp. Fixed Char money overflowing (was getting casted to a short). Fixed inventory gather/sorting.

+++ src/client/character/CharacterStat.java
@@ -225,7 +225,7 @@ public class CharacterStat {
     }
 
     public long getMoney() {
-        return (short) money;
+        return money;
     }
 
     public ExtendSP getExtendSP() {
